.history/requirements_20260428075651.txt
openai>=1.30.1
langchain>=0.2.1
langchain-openai>=0.1.8
langchain-community>=0.2.1
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
tiktoken>=0.7.0
streamlit>=1.35.0
python-dotenv>=1.0.1
scikit-learn>=1.3.0
numpy>=1.24.0
